+++ src/Frontend/Dfe.Complete/Pages/Projects/TaskList/Tasks/HandoverWithDeliveryOfficerTask/HandoverWithDeliveryOfficerTask.cshtml.cs
@@ -1,4 +1,7 @@
+using Dfe.Complete.Application.Projects.Commands.UpdateProject;
+using Dfe.Complete.Application.Projects.Queries.TaskData;
 using Dfe.Complete.Constants;
+using Dfe.Complete.Domain.ValueObjects;
 using Dfe.Complete.Models;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
@@ -19,9 +22,25 @@ namespace Dfe.Complete.Pages.Projects.TaskList.Tasks.HandoverWithDeliveryOfficer
         [BindProperty(Name = "attend-handover-meeting")]
         public bool? AttendHandoverMeeting { get; set; }
 
+        public override async Task<IActionResult> OnGetAsync()
+        {
+            await UpdateCurrentProject(); 
+            await SetEstablishmentAsync();
+            var result = await sender.Send(new GetTaskDataByProjectIdQuery(new ProjectId(Guid.Parse(ProjectId))));
+            if(result.IsSuccess && result.Value != null)
+            {
+                ReviewProjectInformation = result.Value.HandoverReview;
+                MakeNotes = result.Value.HandoverNotes;
+                AttendHandoverMeeting = result.Value.HandoverMeeting;
+                NotApplicable = result.Value.HandoverNotApplicable;
+            }
+            return Page();
+        }
+
         public async Task<IActionResult> OnPost()
         {
-            return Redirect(string.Format(RouteConstants.ProjectHandoverWithDeliveryOfficerTask, ProjectId));
+            _ = await sender.Send(new UpdateHandoverWithDeliveryOfficerCommand(new ProjectId(Guid.Parse(ProjectId)), NotApplicable, ReviewProjectInformation, MakeNotes, AttendHandoverMeeting));
+            return Redirect(string.Format(RouteConstants.ProjectTaskList, ProjectId));
         }
     }
 }