+++ CHANGELOG.md
@@ -20,7 +20,7 @@ Security in case of vulnerabilities.
 
 ---
 
-## Unreleased  
+## Unreleased
 
 ### Added
 - Added Project Significant History Dates `/projects/{projectId}/date-history`
@@ -29,9 +29,32 @@ Security in case of vulnerabilities.
 ### Changed
 
 ### Fixed
+- Fixed cookies page's URL by including query string.
+- Fixed `Local Authority` service support endpoints
+
+See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-07-24.577...main) for everything awaiting release
+
+---
+
+## [1.10.2](https://github.com/DFE-Digital/complete-conversions-transfers-changes/releases/tag/production-2025-07-24.577) - 2025-07-24
+
+### Added
+
+### Changed
+
+### Fixed
+- HOTFIX: antiforgery issue when accepting cookies from Ruby app
+
+See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-07-22.558...production-2025-07-24.577) for everything in the release
+
+---
+
+## [1.10.1](https://github.com/DFE-Digital/complete-conversions-transfers-changes/releases/tag/production-2025-07-22.558) - 2025-07-22
+
+### Fixed
 - HOTFIX: Reports tab now points to exports and dotnet app will handle redirect
 
-See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-07-22.553...main) for everything awaiting release
+See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-07-22.553...production-2025-07-22.558) for everything in the release
 
 ---